server/requirements.txt
attrs==22.2.0
autopep8==2.0.0
blinker==1.5
bson==0.5.10
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
debugpy==1.6.4
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lts==0.0.2
MarkupSafe==2.1.1
mongoengine==0.24.2
mongomock==4.1.2
packaging==22.0
parameterized==0.8.1
pluggy==1.0.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
PySocks==1.7.1
pytest-flask==1.2.0
pytest-mock==3.10.0
python-dateutil==2.8.2
requests==2.28.1
requests-file==1.5.1
sentinels==1.0.0
six==1.16.0
tldextract==3.4.0
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.2
